+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+
+## 1.0.1
+
+* feat: add `isValid` method for checking pricing validity
+* fix: allow pricing model as cace-insensitive
+* fix: data getter default for accessing nullish fields